[inductor] Fix no-xdim reductions
Fixes #103481

Normally triton tensors have shape `[XBLOCK, RBLOCK]`, or some variation where
the lengths are 1 but the number of dimensions is the same. The `no_x_dim`
change in addition to removing the x dimension, also removed the r dimension
from certain values such as the results of reductions and the `xindex` variable.

This fixes those two cases to correctly produce tensors of shape `[1]`,
equivalent to the old shape `[XBLOCK, 1]` with the x-dimension dropped.

ghstack-source-id: 8c46c05c41a5d47daa1f51db06e0e77a16be9739
Pull Request resolved: https://github.com/pytorch/pytorch/pull/103527